move blockhash or number
